refactor components
and its imports used in other files
