+++ src/components/ArticleBox/ArticleBox.tsx
@@ -3,14 +3,14 @@ import { View, TouchableOpacity, Text } from 'react-native';
 
 import { useNavigation } from '@react-navigation/native';
 
-import Thumbnail from '@/components/Thumbnail';
+import Thumbnail from '@/components/Thumbnail/Thumbnail';
 import { Need } from '@/constants/Enum';
 import { remainTime, calcTimeDiff } from '@/helpers/functions/time';
 import { $thumnail_size } from '@/styles/size';
 import { IArticleSumProps } from '@/types/article';
 
 import styles from './ArticleBox.style';
-import GoalBar from './GoalBar';
+import GoalBar from './GoalBar/GoalBar';
 
 function ArticleBox({
   article_id,
