+++ src/components/ArticleBox/GoalBar/GoalBar.tsx
