+++ src/components/InfoContainer/InfoContainer.tsx
