+++ src/components/StringInput/StringInput.tsx
