+++ src/components/Thumbnail/Thumbnail.tsx
