+++ src/components/index.ts
@@ -1,10 +1,10 @@
-export { default as ArticleBox } from './ArticleBox';
-export { default as Button } from './Button';
-export { default as FootTerms } from './FootTerms';
-export { default as HDRBackBtn } from './HDRBackBtn';
-export { default as Header } from './Header';
-export { default as InfoContainer } from './InfoContainer';
-export { default as Profile } from './ProfileBox';
-export { default as StringInput } from './StringInput';
-export { default as Thumbnail } from './Thumbnail';
-export { default as WriteArticle } from './WriteArticle';
+export { default as ArticleBox } from './ArticleBox/ArticleBox';
+export { default as Button } from './Button/Button';
+export { default as FootTerms } from './FootTerms/FootTerms';
+export { default as HDRBackBtn } from './HDRBackBtn/HDRBackBtn';
+export { default as Header } from './Header/Header';
+export { default as InfoContainer } from './InfoContainer/InfoContainer';
+export { default as Profile } from './ProfileBox/ProfileBox';
+export { default as StringInput } from './StringInput/StringInput';
+export { default as Thumbnail } from './Thumbnail/Thumbnail';
+export { default as WriteArticle } from './WriteArticle/WriteArticle';
