+++ src/screens/Article/GoalTopBubble/index.tsx
@@ -3,7 +3,7 @@ import { LayoutChangeEvent } from 'react-native';
 
 import { View } from 'native-base';
 
-import GoalBar from '@/components/ArticleBox/GoalBar';
+import GoalBar from '@/components/ArticleBox/GoalBar/GoalBar';
 import { Need } from '@/constants/Enum';
 import { IParticipantsSummary } from '@/types/article';
